refactor(public-api): move DB queries from web/server into shared repositories
Move ClickHouse and Prisma query functions out of
`web/src/features/public-api/server/` into `packages/shared/src/server/repositories/`
so that data access is co-located with other repository functions and the
public-API handler files are narrowly focused on HTTP concerns.

What moved:
- scores v1/v2 list/count queries → repositories/scores.ts
- scores v3 list query, column constants, filter helpers, cursor helpers → repositories/scores.ts
- traces list/count query → repositories/traces.ts
- observations list/count query → repositories/observations.ts
- daily metrics queries → new repositories/daily-metrics.ts
- createOrFetchDatasetRun → new repositories/dataset-runs.ts
- listScoreConfigs / getScoreConfig → new repositories/score-configs.ts

What stayed in web:
- convertScoreToPublicApi, polymorphicValue, domainToV3 (API-shape mapping)
- createScoreConfig, updateScoreConfig (interleaved with auditLog)
- All HTTP middleware, route factories, auth, rate-limit, tRPC routers

Web files are now thin re-export stubs that preserve all existing import paths.
LANGFUSE_API_CLICKHOUSE_PROPAGATE_OBSERVATIONS_TIME_BOUNDS added to shared env
so the traces repository can read it without importing from web.

Fix: ScoresApiService.generateScoresForPublicApi now applies convertScoreToPublicApi
after fetching so CORRECTION scores in v2 API correctly map longStringValue→stringValue.
